Fjern resten av ulong-gjengen

+++ src/main/kotlin/no/nav/helsearbeidsgiver/inntektsmelding/InntektsmeldingRepository.kt
@@ -76,7 +76,7 @@ class InntektsmeldingRepository(
             filter.navReferanseId?.let { query.andWhere { navReferanseId eq it } }
             filter.fom?.let { query.andWhere { innsendt greaterEq it.tilTidspunktStartOfDay() } }
             filter.tom?.let { query.andWhere { innsendt lessEq it.tilTidspunktEndOfDay() } }
-            filter.fraLoepenr?.let { query.andWhere { InntektsmeldingEntitet.id greater it.toULong() } }
+            filter.fraLoepenr?.let { query.andWhere { InntektsmeldingEntitet.id greater it } }
             query.orderBy(InntektsmeldingEntitet.id, SortOrder.ASC)
             query.limit(MAX_ANTALL_I_RESPONS + 1) // Legg på en, for å kunne sjekke om det faktisk finnes flere enn max antall
             query.map { it.toExposedInntektsmelding() }
